feat: function and app to set query control mode and set control mode

+++ src/openarm/damiao_motor/dm_motor.cpp
@@ -42,7 +42,7 @@ double Motor::get_param(int RID) const {
     return (it != temp_param_dict_.end()) ? it->second : -1;
 }
 
-void Motor::set_temp_param(int RID, int val) { temp_param_dict_[RID] = val; }
+void Motor::set_temp_param(int RID, int val) { temp_param_dict_[RID] = static_cast<double>(val); }
 
 // State update methods
 void Motor::update_state(double q, double dq, double tau, int tmos, int trotor) {
